requirements-lock.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.19
decorator==5.2.1
dill==0.3.8
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.18.0
flake8==7.3.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==3.10
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
jupytext==1.18.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.10.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic-core==2.33.2
pyflakes==3.4.0
pygments==2.19.2
pyparsing==3.2.3
pyright==1.1.408
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.30.0
ruff==0.14.11
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.32.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
triton==3.3.0
trl==0.17.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.14
widgetsnbextension==4.0.15
xxhash==3.5.0
yarl==1.20.1
